requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
crewai==0.80.0
openai
langchain
langchain-openai
python-multipart
python-dotenv
jinja2
aiofiles
